added authentication

+++ src/php/lumen/routes/web.php
@@ -11,23 +11,26 @@
 |
 */
 
-use Illuminate\Http\Request;
-use App\Http\Controllers\MovieController;
-
 
 $router->group(['prefix' => 'api'], function () use ($router) {
     $router->get('movies',  ['uses' => 'MovieController@getAll']);
-
     $router->get('movies/{id}', ['uses' => 'MovieController@getById']);
-
-    $router->post('movies', ['uses' => 'MovieController@create']);
-
-    $router->delete('movies/{id}', ['uses' => 'MovieController@delete']);
-
-    $router->put('movies/{id}', ['uses' => 'MovieController@update']);
+    $router->post('movies', ['middleware' => 'auth', 'uses' => 'MovieController@create']);
+    $router->delete('movies/{id}', ['middleware' => 'auth', 'uses' => 'MovieController@delete']);
+    $router->put('movies/{id}', ['middleware' => 'auth', 'uses' => 'MovieController@update']);
 });
 
 $router->get('/hello', function () use ($router) {
     return 'hello world!';
 });
 
+// !!!! Security Warning!!!!
+// You can uncomment this section for troubleshoothing.
+// However, you must be aware of serious security risks, as it shows your secret key
+// to whoever invokes the API. At least replace SECRET with a string that you know
+// and comment again immediately.
+/*
+$router->get('/phpinfo-SECRET', function () use ($router) {
+    phpinfo();
+});
+*/
